Move u2f components back to app
