+++ app/(auth)/mfa/set/page.tsx
@@ -17,7 +17,7 @@ import { checkSessionFactorValidity, loadSessionById } from "@lib/session";
 import { getSerializableLoginSettings } from "@lib/zitadel";
 import { serverTranslation } from "@i18n/server";
 import { AuthPanel } from "@components/auth/AuthPanel";
-import { ChooseSecondFactorToSetup } from "@components/mfa/ChooseSecondFactorToSetup";
+import { ChooseSecondFactorToSetup } from "@root/app/(auth)/u2f/set/components/ChooseSecondFactorToSetup";
 export async function generateMetadata(): Promise<Metadata> {
   const { t } = await serverTranslation("mfa");
   return { title: t("set.title") };
